requirements.txt
torch==2.6.0
transformers==4.49.0
datasets==3.3.2
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
pyyaml==6.0.2
pytest==8.3.4
pytest-cov==4.1.0
matplotlib==3.10.1
tensorboard==2.19.0
wandb==0.19.7
accelerate==0.26.0
colorama==0.4.6
huggingface_hub==0.29.1
tqdm==4.67.1
peft==0.14.0  # For Parameter-Efficient Fine-Tuning (LoRA)
bitsandbytes==0.45.3  # For quantization
rouge-score==0.1.2  # For generation model evaluation
sentencepiece==0.2.0  # Required by some tokenizers
protobuf==5.26.1  # Compatibility with TensorBoard
safetensors==0.5.3  # For improved model storage
gradio==5.20.0  # For interactive model evaluation

# Required for config optimizer
psutil>=5.9.0
py3nvml>=0.2.7
pyyaml>=6.0